requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
gunicorn==20.1.0 
werkzeug==2.3.0